LanguageEditor fix for saving translation with html
sb65 - 09/11/2013

Nice, but when trying to update a language entry that has html in it,
such as $text_welcome on common/header, then the update is not saved.
This seems to have been reported above as well. This occurs both on
OC1.5.6 and OC1.5.5.

This appears to be because the variable $current_entry is not found by
the str_replace function at line 223 of LanguageEditor.php, I think
because of this line:

$current_entry = '$_[\''.$str.'\']'.$line_spacing."=
'".addslashes($current_var_value)."';";

since the needle text is not found.

Removing the addslashes function appears to fix the problem and I
haven't yet seen issue anywhere else.

@Gatto, I think this is your problem as well. However, the Information,
About Us links are the title of Information pages - and are edited via
Catalog > Information rather than via the language files.

The revised line should be:

$current_entry = '$_[\''.$str.'\']'.$line_spacing."=
'".$current_var_value."';";

http://www.opencart.com/index.php?route=extension/extension/info&extension_id=10791&filter_license=0&filter_download_id=31&page=54